requirements.txt
requests
chevron
pyyaml
rdflib
pandas